impl: source code for translation tool
This is a code using c language, for both the client and the server. The client will connect to the server, and the server will asked the client whether they prefer translation from English to Pig Latin or from Pig Latin to English. The README.pdf serves as a demo of how it works.